pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-analyst-agent"
version = "0.1.0"
description = "Autonomous BI Analyst built with LangGraph and MCP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langgraph>=0.2.0",
  "mcp>=1.0.0",
  "langchain-openai>=0.2.0",
  "pydantic>=2.8.0",
  "psycopg2-binary>=2.9.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
